+++ src/sh1106/sh1106.h
@@ -15,7 +15,7 @@ int sh1106_init(struct sh1106_dev *sh1106);
  * @param data Pointer to data
  * @param n_bytes Number of data bytes to transmit
  */
-int sh1106_send_data(struct sh1106_dev *sh1106, uint8_t *data, uint8_t n_bytes);
+int sh1106_send_data(struct sh1106_dev *sh1106, uint8_t *data, uint16_t n_bytes);
 
 /**
  * @brief Sends a command to the display
